Refactor Thumbnail class so that it does not depend on ExifParserResult.
